refactor(agents): split pdf maxBytesMb tests to own file
