+++ src/agents/tools/pdf-tool.test.ts
@@ -282,179 +282,6 @@ describe("createPdfTool", () => {
     resolveSpy.mockRestore();
   });
 
-  it("rejects when no pdf input provided", async () => {
-    await withConfiguredPdfTool(async (tool) => {
-      await expect(tool.execute("t1", { prompt: "test" })).rejects.toThrow("pdf required");
-    });
-  });
-
-  it("rejects too many PDFs", async () => {
-    await withConfiguredPdfTool(async (tool) => {
-      const manyPdfs = Array.from({ length: 15 }, (_, i) => `/tmp/doc${i}.pdf`);
-      const result = await tool.execute("t1", { prompt: "test", pdfs: manyPdfs });
-      expectFields(result.details, { error: "too_many_pdfs" });
-    });
-  });
-
-  it("rejects invalid maxBytesMb before loading PDFs", async () => {
-    await withConfiguredPdfTool(async (tool) => {
-      const loadSpy = vi.spyOn(webMedia, "loadWebMediaRaw");
-
-      await expect(
-        tool.execute("t1", {
-          prompt: "test",
-          pdf: "/tmp/doc.pdf",
-          maxBytesMb: 0,
-        }),
-      ).rejects.toThrow("maxBytesMb must be greater than 0");
-      expect(loadSpy).not.toHaveBeenCalled();
-    });
-  });
-
-  it("passes validated maxBytesMb to PDF loading", async () => {
-    await withTempPdfAgentDir(async (agentDir) => {
-      const { loadSpy } = await stubPdfToolInfra(agentDir, {
-        provider: "anthropic",
-        input: ["text", "document"],
-      });
-      vi.spyOn(pdfNativeProviders, "anthropicAnalyzePdf").mockResolvedValue("native summary");
-      const cfg = withPdfModel(ANTHROPIC_PDF_MODEL);
-      const tool = requirePdfTool((await loadCreatePdfTool())({ config: cfg, agentDir }));
-
-      await tool.execute("t1", {
-        prompt: "summarize",
-        pdf: "/tmp/doc.pdf",
-        maxBytesMb: "0.5",
-      });
-
-      const [, loadOptions] = firstMockCall(loadSpy, "loadWebMediaRaw");
-      expectFields(loadOptions, { maxBytes: 524_288 });
-      expect(modelAuth.getApiKeyForModel).toHaveBeenCalledWith(
-        expect.objectContaining({ secretSentinels: true }),
-      );
-    });
-  });
-
-  it("clamps pathological maxBytesMb to the cap", async () => {
-    await withTempPdfAgentDir(async (agentDir) => {
-      const { loadSpy } = await stubPdfToolInfra(agentDir, {
-        provider: "anthropic",
-        input: ["text", "document"],
-      });
-      vi.spyOn(pdfNativeProviders, "anthropicAnalyzePdf").mockResolvedValue("native summary");
-      const tool = requirePdfTool(
-        (await loadCreatePdfTool())({ config: withPdfModel(ANTHROPIC_PDF_MODEL), agentDir }),
-      );
-
-      await tool.execute("t1", { prompt: "ocr", pdf: "/tmp/doc.pdf", maxBytesMb: "1000000000" });
-      const [, loadOptions] = firstMockCall(loadSpy, "loadWebMediaRaw");
-      expectFields(loadOptions, { maxBytes: 100 * 1024 * 1024 });
-    });
-  });
-
-  it("real pdf invocation clamps oversized maxBytesMb without schema rejection", async () => {
-    await withTempPdfAgentDir(async (agentDir) => {
-      const workspaceDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-pdf-ws-"));
-      const pdfPath = path.join(workspaceDir, "doc.pdf");
-      await fs.writeFile(pdfPath, FAKE_PDF_MEDIA.buffer);
-      try {
-        const { loadSpy } = await stubPdfToolInfra(agentDir, {
-          mockLoad: false,
-          provider: "anthropic",
-          input: ["text", "document"],
-        });
-        vi.spyOn(pdfNativeProviders, "anthropicAnalyzePdf").mockResolvedValue("native summary");
-        const tool = requirePdfTool(
-          (await loadCreatePdfTool())({
-            config: withPdfModel(ANTHROPIC_PDF_MODEL),
-            agentDir,
-            workspaceDir,
-          }),
-        );
-
-        const result = await tool.execute("t1", {
-          prompt: "summarize",
-          pdf: pdfPath,
-          maxBytesMb: "1000000000",
-        });
-
-        expect(result.content).toEqual([{ type: "text", text: "native summary" }]);
-        const [loadRef, loadOptions] = firstMockCall(loadSpy, "loadWebMediaRaw");
-        expect(loadRef).toBe(pdfPath);
-        expectFields(loadOptions, { maxBytes: 100 * 1024 * 1024 });
-      } finally {
-        await fs.rm(workspaceDir, { recursive: true, force: true });
-      }
-    });
-  });
-
-  it("uses configuredMaxBytesMb when omitted and passes below-cap through", async () => {
-    await withTempPdfAgentDir(async (agentDir) => {
-      const { loadSpy } = await stubPdfToolInfra(agentDir, {
-        provider: "anthropic",
-        input: ["text", "document"],
-      });
-      vi.spyOn(pdfNativeProviders, "anthropicAnalyzePdf").mockResolvedValue("native summary");
-
-      // Configured fallback: no model-supplied maxBytesMb
-      const cfg = {
-        ...withPdfModel(ANTHROPIC_PDF_MODEL),
-        agents: {
-          defaults: { pdfMaxMb: 50, pdfModel: { primary: ANTHROPIC_PDF_MODEL } },
-        },
-      } as OpenClawConfig;
-      const tool = requirePdfTool((await loadCreatePdfTool())({ config: cfg, agentDir }));
-      await tool.execute("t1", { prompt: "ocr", pdf: "/tmp/doc.pdf" });
-      expectFields(firstMockCall(loadSpy, "loadWebMediaRaw")[1], {
-        maxBytes: 50 * 1024 * 1024,
-      });
-
-      loadSpy.mockClear();
-      await tool.execute("t1", { prompt: "ocr", pdf: "/tmp/doc.pdf", maxBytesMb: "50" });
-      expectFields(firstMockCall(loadSpy, "loadWebMediaRaw")[1], {
-        maxBytes: 50 * 1024 * 1024,
-      });
-    });
-  });
-
-  it("respects fsPolicy.workspaceOnly for non-sandbox pdf paths", async () => {
-    await withTempPdfAgentDir(async (agentDir) => {
-      const workspaceDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-pdf-ws-"));
-      const outsideDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-pdf-out-"));
-      try {
-        const cfg = withPdfModel(ANTHROPIC_PDF_MODEL);
-        const tool = requirePdfTool(
-          (await loadCreatePdfTool())({
-            config: cfg,
-            agentDir,
-            workspaceDir,
-            fsPolicy: { workspaceOnly: true },
-          }),
-        );
-
-        const outsidePdf = path.join(outsideDir, "secret.pdf");
-        await fs.writeFile(outsidePdf, "%PDF-1.4 fake");
-
-        await expect(tool.execute("t1", { prompt: "test", pdf: outsidePdf })).rejects.toThrow(
-          /not under an allowed directory/i,
-        );
-      } finally {
-        await fs.rm(workspaceDir, { recursive: true, force: true });
-        await fs.rm(outsideDir, { recursive: true, force: true });
-      }
-    });
-  });
-
-  it("rejects unsupported scheme references", async () => {
-    await withConfiguredPdfTool(async (tool) => {
-      const result = await tool.execute("t1", {
-        prompt: "test",
-        pdf: "ftp://example.com/doc.pdf",
-      });
-      expectFields(result.details, { error: "unsupported_pdf_reference" });
-    });
-  });
-
   it("resolves media://inbound PDF refs", async () => {
     await withManagedInboundPdf(async ({ mediaId }) => {
       await withTempPdfAgentDir(async (agentDir) => {
